+++ CmfSeoBundle.php
@@ -16,7 +16,6 @@ use Doctrine\Bundle\DoctrineBundle\DependencyInjection\Compiler\DoctrineOrmMappi
 
 use Symfony\Cmf\Bundle\SeoBundle\DependencyInjection\Compiler\RegisterBestMatcherPass;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
-use Symfony\Component\DependencyInjection\Compiler\PassConfig;
 use Symfony\Component\HttpKernel\Bundle\Bundle;
 
 use Symfony\Cmf\Bundle\SeoBundle\DependencyInjection\Compiler\RegisterExtractorsPass;
@@ -53,7 +52,8 @@ class CmfSeoBundle extends Bundle
                     realpath(__DIR__ . '/Resources/config/doctrine-phpcr') => 'Symfony\Cmf\Bundle\SeoBundle\Doctrine\Phpcr'
                 ),
                 array('cmf_seo.dynamic.persistence.phpcr.manager_name'),
-                'cmf_seo.backend_type_phpcr'
+                'cmf_seo.backend_type_phpcr',
+                array('CmfSeoBundle' => 'Symfony\Cmf\Bundle\SeoBundle\Doctrine\Phpcr')
             )
         );
     }
@@ -79,7 +79,8 @@ class CmfSeoBundle extends Bundle
                     realpath(__DIR__ . '/Resources/config/doctrine-model') => 'Symfony\Cmf\Bundle\SeoBundle\Model',
                 ),
                 array('cmf_routing.dynamic.persistence.orm.manager_name'),
-                'cmf_routing.backend_type_orm'
+                'cmf_routing.backend_type_orm',
+                array('CmfSeoBundle' => 'Symfony\Cmf\Bundle\SeoBundle\Model')
             )
         );
     }