+++ DependencyInjection/CmfSeoExtension.php
@@ -11,7 +11,6 @@
 
 namespace Symfony\Cmf\Bundle\SeoBundle\DependencyInjection;
 
-use Symfony\Component\Config\Definition\Exception\InvalidConfigurationException;
 use Symfony\Component\Config\FileLocator;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Loader\XmlFileLoader;
@@ -25,6 +24,11 @@ use Symfony\Cmf\Bundle\RoutingBundle\Routing\DynamicRouter;
  */
 class CmfSeoExtension extends Extension
 {
+    /**
+     * @var string
+     */
+    private $defaultAlternateLocaleProviderId;
+
     /**
      * {@inheritDoc}
      */
@@ -57,6 +61,8 @@ class CmfSeoExtension extends Extension
                 $config['persistence']['phpcr']['manager_name']
             );
             $sonataBundles[] = 'SonataDoctrinePHPCRAdminBundle';
+
+            $this->loadPhpcr($config['persistence']['phpcr'], $loader, $container);
         }
 
         if ($this->isConfigEnabled($container, $config['persistence']['orm'])) {
@@ -72,8 +78,8 @@ class CmfSeoExtension extends Extension
             $this->loadSonataAdmin($config['sonata_admin_extension'], $loader, $container, $sonataBundles);
         }
 
-        if ($this->isConfigEnabled($container, $config['error_handling'])) {
-            $this->loadErrorHandling($config['error_handling'], $loader, $container);
+        if ($this->isConfigEnabled($container, $config['alternate_locale'])) {
+            $this->loadAlternateLocaleProvider($config['alternate_locale'], $container);
         }
     }
 
@@ -142,12 +148,44 @@ class CmfSeoExtension extends Extension
      * @param XmlFileLoader $loader
      * @param ContainerBuilder $container
      */
-    private function loadErrorHandling($config, XmlFileLoader $loader, ContainerBuilder $container)
+    private function loadPhpcr($config, XmlFileLoader $loader, ContainerBuilder $container)
     {
-        if ($container->hasParameter($this->getAlias().'.backend_type_phpcr')
-            && $container->getParameter($this->getAlias().'.backend_type_phpcr')
-        ) {
+        $bundles = $container->getParameter('kernel.bundles');
+        if (isset($bundles['CmfRoutingBundle'])) {
+            $loader->load('phpcr-alternate-locale.xml');
+            if (!$this->defaultAlternateLocaleProviderId) {
+                $this->defaultAlternateLocaleProviderId = 'cmf_seo.alternate_locale.provider_phpcr';
+            }
+
             $loader->load('matcher_phpcr.xml');
+        }
+    }
+
+    /**
+     * When setting a custom alternate locale provider with its id, this one will
+     * be injected to the content listener.
+     *
+     * When using phpcr-odm a default provider will be set, when choosing no own one.
+     *
+     * @param array             $config
+     * @param ContainerBuilder $container
+     */
+    private function loadAlternateLocaleProvider($config, ContainerBuilder $container)
+    {
+
+        $alternateLocaleProvider = empty($config['provider_id'])
+            ? $this->defaultAlternateLocaleProviderId
+            : $config['provider_id'];
+
+
+        if ($alternateLocaleProvider) {
+            $definition = $container->getDefinition('cmf_seo.event_listener.seo_content');
+            $definition
+                ->addMethodCall(
+                    'setAlternateLocaleProvider',
+                    array($container->getDefinition($alternateLocaleProvider))
+                )
+            ;
         }
     }
 }